requirements.txt
python-telegram-bot==21.7
entsoe-py==0.6.11
torch==2.5.1
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.0
python-dotenv==1.0.1
apscheduler==3.10.4
